+++ packages/jest-message-util/src/__tests__/messages.test.ts
@@ -413,3 +413,39 @@ it('getTopFrame should return a path for mjs files', () => {
 
   expect(frame!.file).toBe(expectedFile);
 });
+
+it('should return the error cause if there is one', () => {
+  const error = new Error('Test exception');
+  // TODO pass `cause` to the `Error` constructor when lowest supported Node version is 16.9.0 and above
+  // See https://github.com/nodejs/node/blob/main/doc/changelogs/CHANGELOG_V16.md#error-cause
+  error.cause = new Error('Cause Error');
+  const message = formatExecError(
+    error,
+    {
+      rootDir: '',
+      testMatch: [],
+    },
+    {
+      noStackTrace: false,
+    },
+  );
+  expect(message).toMatchSnapshot();
+});
+
+it('should print all the errors contained in an AggregateError', () => {
+  const error = new AggregateError([
+    new Error('error 1'),
+    new Error('error 2'),
+  ]);
+  const message = formatExecError(
+    error,
+    {
+      rootDir: '',
+      testMatch: [],
+    },
+    {
+      noStackTrace: false,
+    },
+  );
+  console.log(message);
+});